Denormalizes the db_user table; simplifies types
After further consideration it turns out that having a separate user
table didn't really make sense. It led to a number of problems:
1) It made it too easy for the number of available users per host and
available databases per host to get out of sync
2) It made assigning and releasing a test database a compound operation
which makes introduces the possibility of partial failures. Now
assigning and releasing a test database is an "atomic operation".
3) Allowing different users to be assigned to different databases
didn't actually gain us anything except for the above problems.

For that reason the db_user table has been moved into the db_state table
(which has since been renamed to test_db which is a more fitting name).
The user_assignment table is obselete and has been removed.

As a consequence, the interface of the manager backend has been
considerably simplified (get_user and assign_user are no longer
necessary). In addition there is no longer a  need for phantom types;
all the remaining operations in the backend database are valid for all
types of test database (i.e. irrespective if the database is in use,
fresh, open, etc.). This has simplified the type representation of
database handles (although to the db manager itself the database handle
type remains abstract).

Further, the administrative interface has been updated to require a new
user to be added when a new test database is added thus ensuring that
problem 1 above is avoided.

Coinciding with the above change in the test_db table the db_hosts table
has also been renamed to db_host to keep things consistent.